+++ src/index.js
@@ -3,9 +3,44 @@ import './style.css';
 import sync from './assets/icons/sync.png';
 import downLeft from './assets/icons/down_left.png';
 import { tasks } from './modules/displayTask.js';
+import menuIcon from './assets/icons/menuIcon.png';
 
+const todoContaiter = document.querySelector('.todoContaiter');
 const titleContainer = document.querySelector('.title-list');
 const descriptionContainer = document.querySelector('.description-list');
+
+const tasksArray = [
+  {
+    index: 1,
+    description: 'first task',
+    completed: false,
+  },
+  {
+    index: 2,
+    description: 'second task',
+    completed: false,
+  },
+  {
+    index: 3,
+    description: 'third task',
+    completed: false,
+  },
+];
+
+
+tasksArray.forEach((task) => {
+  const taskHtml = `<ul class="task-list">
+  <li class="task-list-item">
+    <input type="checkbox" name="check-task" id="check"/>
+  </li>
+  <li class="task-list-item">${task.description}</li>
+  </li>
+  <li class="task-list-item">
+    <img class="icon" src=${menuIcon} />
+  </li>
+</ul>`;
+  todoContaiter.innerHTML += taskHtml;
+});
 
 const titleHtml = `
   <li class="title-list-item">
@@ -31,5 +66,3 @@ const descriptionHtml = `
   </li>
 `;
 descriptionContainer.innerHTML = descriptionHtml;
-
-tasks();
